backend/requirement.txt
fastapi==0.100.0
uvicorn==0.22.0
SQLAlchemy==2.0.13
psycopg2-binary==2.9.6
python-dotenv==1.0.0
alembic==1.11.1
pytest==7.4.0
httpx==0.24.1
requests==2.31.0
python-jose[cryptography]==3.3.0
PyJWT==2.1.0